Move `DiscoveredFiles` into `Files`
